Moved all components *.js files to corresponding subdirectories, moved their stories there. Created index.js file for each component for pretty import. Fixed imports from components
