+++ components/FilterBar/FilterBar.js
@@ -1,5 +1,5 @@
 import { useRecoilState } from 'recoil'
-import InputContainer from '../InputContainer/InputContainer'
+import InputContainer from '../InputContainer'
 import { formFiltersState } from '../Catalog'
 
 const FilterBar = () => {
